Split app.go, main_test.go, get rid of `seqs`
Problems:
1. Large code files make it almost impossible to understand the codebase
2. Usage of global `seqs` contains a potential trouble when executing
   tests in parallel

Solution: break up `app.go` and `main_test.go` files into multiple
smaller files. Additionally, get rid of global `seqs` channel and replace
it with per-application counter under a mutex.
