+++ src/app/libp2p_helper/src/libp2p_helper/config_msg_test.go
@@ -1,0 +1,379 @@
+package main
+
+import (
+	crand "crypto/rand"
+	"errors"
+	"io/ioutil"
+	"testing"
+	"time"
+
+	capnp "capnproto.org/go/capnp/v3"
+	"github.com/libp2p/go-libp2p-core/crypto"
+	peer "github.com/libp2p/go-libp2p-core/peer"
+	ma "github.com/multiformats/go-multiaddr"
+	ipc "libp2p_ipc"
+
+	"github.com/stretchr/testify/require"
+)
+
+func maToStringList(vs []ma.Multiaddr) []string {
+	vsm := make([]string, len(vs))
+	for i, v := range vs {
+		vsm[i] = v.String()
+	}
+	return vsm
+}
+
+func beginAdvertisingSendAndCheck(t *testing.T, app *app) {
+	_, seg, err := capnp.NewMessage(capnp.SingleSegment(nil))
+	require.NoError(t, err)
+	m, err := ipc.NewRootLibp2pHelperInterface_BeginAdvertising_Request(seg)
+	require.NoError(t, err)
+	var rpcSeqno uint64 = 123
+	resMsg := app.handleBeginAdvertising(rpcSeqno, m)
+	seqno, respSuccess := checkRpcResponseSuccess(t, resMsg)
+	require.Equal(t, seqno, rpcSeqno)
+	require.True(t, respSuccess.HasBeginAdvertising())
+	_, err = respSuccess.BeginAdvertising()
+	require.NoError(t, err)
+}
+
+/*
+  Tests for BeginAdvertising message and
+  various discovery mechanisms
+*/
+
+func TestDHTDiscovery_TwoNodes(t *testing.T) {
+	appA, _ := newTestApp(t, nil, true)
+	appA.NoMDNS = true
+
+	appAInfos, err := addrInfos(appA.P2p.Host)
+	require.NoError(t, err)
+
+	appB, _ := newTestApp(t, appAInfos, true)
+	appB.AddedPeers = appAInfos
+	appB.NoMDNS = true
+
+	// begin appB and appA's DHT advertising
+	beginAdvertisingSendAndCheck(t, appB)
+	beginAdvertisingSendAndCheck(t, appA)
+
+	time.Sleep(time.Second)
+}
+
+func TestDHTDiscovery_ThreeNodes(t *testing.T) {
+	appA, _ := newTestApp(t, nil, true)
+	appA.NoMDNS = true
+
+	appAInfos, err := addrInfos(appA.P2p.Host)
+	require.NoError(t, err)
+
+	appB, _ := newTestApp(t, appAInfos, true)
+	appB.NoMDNS = true
+
+	err = appB.P2p.Host.Connect(appB.Ctx, appAInfos[0])
+	require.NoError(t, err)
+
+	appC, _ := newTestApp(t, appAInfos, true)
+	appC.NoMDNS = true
+
+	err = appC.P2p.Host.Connect(appC.Ctx, appAInfos[0])
+	require.NoError(t, err)
+
+	time.Sleep(time.Second)
+
+	// begin appB and appC's DHT advertising
+	beginAdvertisingSendAndCheck(t, appB)
+	beginAdvertisingSendAndCheck(t, appC)
+
+	done := make(chan struct{})
+
+	go func() {
+		for {
+			// check if peerB knows about peerC
+			addrs := appB.P2p.Host.Peerstore().Addrs(appC.P2p.Host.ID())
+			if len(addrs) != 0 {
+				// send a stream message
+				// then exit
+				close(done)
+				return
+			}
+			time.Sleep(time.Millisecond * 100)
+		}
+	}()
+
+	select {
+	case <-time.After(testTimeout):
+		t.Fatal("B did not discover C via DHT")
+	case <-done:
+	}
+
+	time.Sleep(time.Second)
+}
+
+func TestMDNSDiscovery(t *testing.T) {
+	appA, appAPort := newTestApp(t, nil, true)
+	appA.NoDHT = true
+
+	appB, appBPort := newTestApp(t, nil, true)
+	appB.NoDHT = true
+
+	t.Logf("Using libp2p ports: %d and %d", appAPort, appBPort)
+
+	// begin appA and appB's mDNS advertising
+	beginAdvertisingSendAndCheck(t, appB)
+	beginAdvertisingSendAndCheck(t, appA)
+
+	done := make(chan struct{})
+
+	go func() {
+		for {
+			// check if peerB knows about peerA
+			addrs := appB.P2p.Host.Peerstore().Addrs(appA.P2p.Host.ID())
+			if len(addrs) != 0 {
+				close(done)
+				return
+			}
+			time.Sleep(time.Millisecond * 100)
+		}
+	}()
+
+	select {
+	case <-time.After(testTimeout):
+		t.Fatal("B did not discover A via mDNS")
+	case <-done:
+	}
+
+	time.Sleep(time.Second * 3)
+}
+
+func TestConfigure(t *testing.T) {
+	testApp := newApp()
+
+	dir, err := ioutil.TempDir("", "mina_test_*")
+	require.NoError(t, err)
+
+	key, _, err := crypto.GenerateEd25519Key(crand.Reader)
+	require.NoError(t, err)
+	keyBytes, err := key.Bytes()
+	require.NoError(t, err)
+
+	external := "/ip4/0.0.0.0/tcp/7000"
+	self := "/ip4/127.0.0.1/tcp/7000"
+
+	_, seg, err := capnp.NewMessage(capnp.SingleSegment(nil))
+	require.NoError(t, err)
+	m, err := ipc.NewRootLibp2pHelperInterface_Configure_Request(seg)
+	require.NoError(t, err)
+	c, err := m.NewConfig()
+	require.NoError(t, err)
+
+	require.NoError(t, c.SetStatedir(dir))
+	require.NoError(t, c.SetPrivateKey(keyBytes))
+	require.NoError(t, c.SetNetworkId(string(testProtocol)))
+	lon, err := c.NewListenOn(1)
+	require.NoError(t, err)
+	require.NoError(t, lon.At(0).SetRepresentation(self))
+	err = multiaddrListForeach(lon, func(v string) error {
+		if v != self {
+			return errors.New("Failed to iterate over freshly created listenOn list")
+		}
+		return nil
+	})
+	require.NoError(t, err)
+	require.True(t, lon.Len() == 1)
+	c.SetMetricsPort(0)
+	ema, err := c.NewExternalMultiaddr()
+	require.NoError(t, err)
+	require.NoError(t, ema.SetRepresentation(external))
+	c.SetUnsafeNoTrustIp(false)
+	c.SetFlood(false)
+	c.SetPeerExchange(false)
+	_, err = c.NewDirectPeers(0)
+	require.NoError(t, err)
+	_, err = c.NewSeedPeers(0)
+	require.NoError(t, err)
+	c.SetMaxConnections(0)
+	c.SetValidationQueueSize(16)
+	c.SetMinaPeerExchange(false)
+
+	gc, err := c.NewGatingConfig()
+	require.NoError(t, err)
+	_, err = gc.NewBannedIps(0)
+	require.NoError(t, err)
+	_, err = gc.NewBannedPeerIds(0)
+	require.NoError(t, err)
+	_, err = gc.NewTrustedIps(0)
+	require.NoError(t, err)
+	_, err = gc.NewTrustedPeerIds(0)
+	require.NoError(t, err)
+	gc.SetIsolate(false)
+
+	resMsg := testApp.handleConfigure(239, m)
+	require.NoError(t, err)
+	seqno, respSuccess := checkRpcResponseSuccess(t, resMsg)
+	require.Equal(t, seqno, uint64(239))
+	require.True(t, respSuccess.HasConfigure())
+	_, err = respSuccess.Configure()
+	require.NoError(t, err)
+}
+
+func TestGenerateKeypair(t *testing.T) {
+	_, seg, err := capnp.NewMessage(capnp.SingleSegment(nil))
+	require.NoError(t, err)
+	m, err := ipc.NewRootLibp2pHelperInterface_GenerateKeypair_Request(seg)
+	require.NoError(t, err)
+
+	testApp, _ := newTestApp(t, nil, true)
+	resMsg := testApp.handleGenerateKeypair(7839, m)
+	require.NoError(t, err)
+	seqno, respSuccess := checkRpcResponseSuccess(t, resMsg)
+	require.Equal(t, seqno, uint64(7839))
+	require.True(t, respSuccess.HasGenerateKeypair())
+	res, err := respSuccess.GenerateKeypair()
+	require.NoError(t, err)
+	r, err := res.Result()
+	require.NoError(t, err)
+
+	sk, err := r.PrivateKey()
+	require.NoError(t, err)
+	pk, err := r.PublicKey()
+	require.NoError(t, err)
+	pid, err := r.PeerId()
+	require.NoError(t, err)
+	peerId, err := pid.Id()
+	require.NoError(t, err)
+
+	// extra 4 bytes due to key type byte + protobuf encoding
+	require.Equal(t, len(sk), 68)
+	// extra 3 bytes due to key type byte + protobuf encoding
+	require.Equal(t, len(pk), 36)
+	require.Greater(t, len(peerId), 0)
+}
+
+func TestGetListeningAddrs(t *testing.T) {
+	testApp, _ := newTestApp(t, nil, true)
+	_, seg, err := capnp.NewMessage(capnp.SingleSegment(nil))
+	require.NoError(t, err)
+	m, err := ipc.NewRootLibp2pHelperInterface_GetListeningAddrs_Request(seg)
+	require.NoError(t, err)
+	var mRpcSeqno uint64 = 1024
+	resMsg := testApp.handleGetListeningAddrs(mRpcSeqno, m)
+	seqno, respSuccess := checkRpcResponseSuccess(t, resMsg)
+	require.Equal(t, seqno, mRpcSeqno)
+	require.True(t, respSuccess.HasGetListeningAddrs())
+	ls, err := respSuccess.GetListeningAddrs()
+	require.NoError(t, err)
+	addrsL, err := ls.Result()
+	res, err := readMultiaddrList(addrsL)
+	require.NoError(t, err)
+	require.Equal(t, maToStringList(testApp.P2p.Host.Addrs()), res)
+}
+
+func TestListen(t *testing.T) {
+	addrStr := "/ip4/127.0.0.2/tcp/8000"
+
+	testApp, _ := newTestApp(t, nil, true)
+
+	_, seg, err := capnp.NewMessage(capnp.SingleSegment(nil))
+	require.NoError(t, err)
+	m, err := ipc.NewRootLibp2pHelperInterface_Listen_Request(seg)
+	require.NoError(t, err)
+	require.NoError(t, m.SetIface(addrStr))
+
+	resMsg := testApp.handleListen(1239, m)
+	require.NoError(t, err)
+	seqno, respSuccess := checkRpcResponseSuccess(t, resMsg)
+	require.Equal(t, seqno, uint64(1239))
+	require.True(t, respSuccess.HasListen())
+	lresp, err := respSuccess.Listen()
+	require.NoError(t, err)
+	l, err := lresp.Result()
+	require.NoError(t, err)
+	found := false
+	err = multiaddrListForeach(l, func(v string) error {
+		if v == addrStr {
+			found = true
+		}
+		return nil
+	})
+	require.NoError(t, err)
+	require.True(t, found)
+}
+
+func TestSetGatingConfig(t *testing.T) {
+	testApp, _ := newTestApp(t, nil, true)
+
+	allowedID := "12D3KooWJDGPa2hiYCJ2o7XPqEq2tjrWpFJzqa4dy538Gfs7Vn2r"
+	allowedMultiaddr, err := ma.NewMultiaddr("/ip4/7.8.9.0/tcp/7000")
+	require.NoError(t, err)
+
+	bannedID := "12D3KooWGnQ4vat8EybAeFEK3jk78vmwDu9qMhZzcyQBPb16VCnS"
+	bannedMultiaddr, err := ma.NewMultiaddr("/ip4/1.2.3.4/tcp/7000")
+	require.NoError(t, err)
+
+	_, seg, err := capnp.NewMessage(capnp.SingleSegment(nil))
+	require.NoError(t, err)
+	m, err := ipc.NewRootLibp2pHelperInterface_SetGatingConfig_Request(seg)
+	require.NoError(t, err)
+
+	gc, err := m.NewGatingConfig()
+	require.NoError(t, err)
+	bIps, err := gc.NewBannedIps(1)
+	require.NoError(t, err)
+	bPids, err := gc.NewBannedPeerIds(1)
+	require.NoError(t, err)
+	tIps, err := gc.NewTrustedIps(1)
+	require.NoError(t, err)
+	tPids, err := gc.NewTrustedPeerIds(1)
+	require.NoError(t, err)
+	require.NoError(t, bIps.Set(0, "1.2.3.4"))
+	require.NoError(t, bPids.At(0).SetId(bannedID))
+	require.NoError(t, tIps.Set(0, "7.8.9.0"))
+	require.NoError(t, tPids.At(0).SetId(allowedID))
+	gc.SetIsolate(false)
+
+	var mRpcSeqno uint64 = 2003
+	resMsg := testApp.handleSetGatingConfig(mRpcSeqno, m)
+	seqno, respSuccess := checkRpcResponseSuccess(t, resMsg)
+	require.Equal(t, seqno, mRpcSeqno)
+	require.True(t, respSuccess.HasSetGatingConfig())
+	_, err = respSuccess.SetGatingConfig()
+	require.NoError(t, err)
+
+	ok := testApp.P2p.GatingState.InterceptPeerDial(peer.ID(bannedID))
+	require.False(t, ok)
+
+	ok = testApp.P2p.GatingState.InterceptPeerDial(peer.ID(allowedID))
+	require.True(t, ok)
+
+	ok = testApp.P2p.GatingState.InterceptAddrDial(peer.ID(bannedID), bannedMultiaddr)
+	require.False(t, ok)
+
+	ok = testApp.P2p.GatingState.InterceptAddrDial(peer.ID(bannedID), allowedMultiaddr)
+	require.False(t, ok)
+
+	ok = testApp.P2p.GatingState.InterceptAddrDial(peer.ID(allowedID), allowedMultiaddr)
+	require.True(t, ok)
+}
+
+func TestSetNodeStatus(t *testing.T) {
+	testApp, _ := newTestApp(t, nil, true)
+
+	_, seg, err := capnp.NewMessage(capnp.SingleSegment(nil))
+	require.NoError(t, err)
+	m, err := ipc.NewRootLibp2pHelperInterface_SetNodeStatus_Request(seg)
+	require.NoError(t, err)
+	testStatus := []byte("test_node_status")
+	require.NoError(t, m.SetStatus(testStatus))
+
+	resMsg := testApp.handleSetNodeStatus(11239, m)
+	require.NoError(t, err)
+	seqno, respSuccess := checkRpcResponseSuccess(t, resMsg)
+	require.Equal(t, seqno, uint64(11239))
+	require.True(t, respSuccess.HasSetNodeStatus())
+	_, err = respSuccess.SetNodeStatus()
+	require.NoError(t, err)
+
+	require.Equal(t, testStatus, testApp.P2p.NodeStatus)
+}
